Finally fix Audio
The source data is in Big-Endian (MSB samples), and the scummvm
decoder used those as-is.   My SDL2 doesn't want to play AUDIO_S16MSB
samples, therefore it sounded like noise with faint hint of the
intended audio embedded in it.

Now the sample generator defaults to little-endian samples by default.