sandec: perf optimization
drastically reduce the number of io->read() calls:  perf report showed
that >70% of the execution time of sanplay was spent in the ->read()
callback to fetch new data.
Rewrite the FRME handler to cache a full FRME in RAM; ANIMv2 has
a field which mentions the maximum size of all FRME objects in the
stream, so we can allocate a suitably sized buffer before decoding
the first frame.

reduces pure render time of OP_CR.SAN file (2047 FRME objects) from 3244ms
down to 540ms, a 6x speedup.

Signed-off-by: Manuel Lauss <manuel.lauss@gmail.com>