sandec: fix Shadows Of The Empire videos
SOTE videos are 640x272 px codec47 videos, which have a constant offset
of 60 from the top.
The comp2 handler did not take this into account for the 2 auxbuffers,
and on sequence reset, only the parts of the auxbuffers specified by
the frame dimensions are cleared, otherwise you'd get a pink 640x60
strip on the L00LOGO.SAN file.
The code is still not correct wrt. handling of "left" offset though,
but will be adressed when I encounter a SAN video that has it set
to anything else than zero.

Outlaws and COMI still work fine after this since they set left and
top to zero on all frames.