sandec: fix codec37 assumptions
Codec37 operates on 2 buffers, which need to be considered private
to the codec only, and must not be touched by other codecs during
playback (codec1 is often used in between c37 frames).

change codec37 to work on buf1/buf2 exclusively and copy the decoded
image to buf0 at the end, in case another codec wants to operate on
it in subsequent FOBJs.  This fixes an issue only observed in the
Full Throttle "dazed.san" file, where the face would only consist
of eyes since the rest of it has been cleared by codec1 FOBJs
in between.

Codecs 47 and 48 (which is derived from c37) seem to expect the
previous frame not to be untouched.

Found by looking at the throttle.exe file with Ghidra.