sanplay: default-enable frame interpolation
Esp. in MotS the video far smoother, esp. the opening crawl and
all the camera panning.

In Outlaws it seems mixed: generally the scrolling is softer and
the "jerkiness" of the source material is almost gone, BUT, when
characters are talking, their mouths have some noticeable "ghosts".

But the good outweighs the bad IMO, so I'll leave it default on
in the player.