sanplay: perfect A/V sync
This change achieves perfect A/V sync on my test machines:
In Outlaws' IN_SHB.SAN (Level 1 Intro) Movie, the sound of the
shovel entering the ground (at around 05:00 minutes of playback)
perfectly matches up with the video.
At the cost of lots of CPU usage though since it's basically now
busy-waiting for the timeline to reach the present timestamp of the
most recently decoded frame.  Will fix that later.