sandec: rewrite codec45 based on disassembly and disable it
It does not change anything though:  the datastream keeps xoff and
yoff outside the visible buffer area for 99.9% of the time so it
does nothing.  With the code a bit adjusted to merge xoff/yoff into
an offset from destination base, the codec pretty much produces
line-artifacts in the initial iris-out shots.

The assembly also references variables in its inner loop which are
only ever read by this codec, but never written to anywhere else
in the code (at least according to Ghidra).

I don't think this ever worked properly and it's only used in the
Rebel Assault 2 "LEVxx/xxRETRY*.SAN" videos during the first few
frames with the iris-out effect.