sandec: STOR/FTCH as implemented in RA1/2
Implement STOR/FTCH like it is done in LucasArts games:
On STOR, the next following FOBJ is cached as-is in an aux buffer,
and replayed on FTCH.
This way we can also take the FTCH offsets into account by adding
them to the left/top values of the stored FOBJ.
This with codec21 update corrects the playback of e.g. RA1 L8PLAY.ANM,
the cockpit (a STORed codec21 object) is now correctly rendered
on top of the terrain+walker.
For the full-buffer codecs37/47/48 we can decode the image as-is
store that in the STOR buffer as a codec20 object, and decode
it like a normal FOBJ.