sandec: PSAD/iMUS streamsync fix
Introduce a required audio fragment size for processing/mixing the
PSAD/iMUS streams.  The size is determined by the ANIM framerate,
and is the minimum amount of bytes required to produce during every
processed FRME to get click-free output.
Incidentally, this also fixes the missing synchronization of the
multiple substreams.
At least the Full Throttle and The Dig intros have the speech
lipsynced to the video.