sandec: fix endless loop mixing tracks without new data
Some RA1 videos, e.g. L8PLAY.ANM, are meant to be continued with
another one, and after the last FRME, some audio tracks still have
more data to receive before being done.
This causes an endless loop waiting for the number of active tracks
to finish.  Rather than waiting for that, end the loop when
the mixer function was not able to make any progress any more.