add flag to display whole ANIMv1 buffer
For ANIMv1, by default, only the upper-left 320x200 area is queued
to the player since this is where most videos have their content.
However the main gameplay videos have content on the whole 384x242
buffer, and this adds the ability to toggle the display of all or
the limited viewport on a per frame basis.

Press "L" in the player window to toggle.