sandec: fix 800x600 BL16 videos
The jonesopn_8.snm from Indiana Jones and the Infernal Machine showed
various artifacts when playing back.
It turns out that opcodes 1-4 were the ones causing issues.
These opcodes are entries into the c47_mv table, where an offset
into the diff buffers is generated.
For width 800, these 4 pairs slightly overflow an int16, turning a
large negative value into a large positive one.  In anticipation of
this, I made the calculated motion vector an int32, keeping the
large negative value.  It turns out that the original smush decoder
relied on this overflow and actively exploited it in the jonesopn_8.snm video.

The original code precomputed a all possible mv offsets into an int16
table.  Here we calculate the offset the same way into an int16,
then store it in an int32 multiplied by 2 (due to 16bit pixels, instead
of the 8bit pixels assumed by c47).