sandec: codec48: reimplement the block scaling of the JKM decoder
Implement the 1x1->8x8 and 2x2->8x8 scaling like the JKM decoder does:
by first creating a 4x4 buffer using the source data and various
neighboring pixels and using the interpolation table to create
intermedia pixels.  Each of these 16 pixels is then blown up to 2x2
to get the 8x8 result.

This is necessary as it fixes a few artifacts in the JKM S1L1OCS.SAN
video between frames 25 and 50 (white blocks in the middle of the frame)