sandec: reimplement XPAL algorithm like DOS players
Implement the XPAL algorithm like the original SMUSH engines.
This gets rid of the visual "pulses" in the Outlaws RAE.SAN
fadeout at the end and makes is smooth and seamless.
Also gets rid of a few color banding issues in The Dig, noticeably
the night sky of the Shuttle launch countdown scene is for more
varied.