sandec: codec48: fix Making Magic video artifacts
Making Magic makes use of flag bit 0, which when set, prevents the
swapping of the 2 work buffers on even sequence numbers.

It also makes use of flag 0x10, which when set, indicates to create
an interpolated frame from the 2 buffers and use this one as result
instead.  Frame with this bit are always followed up with a frame
of compression type 6, which will transfer the decoded result of the
previous frame.

It's still missing handling of flag 0x02, which does a mask blit,
treating color 0 as transparent.