sandec: Full Throttle needs faster playback.
The Full Throttle videos request 10fps, but the rate at which the
audio data comes in per FRME is not enough for 100ms.  It starts
to work when playback is sped up to 10.2fps.

For SAN videos with framerate slower than 11 fps it is adjusted to
10,5 fps because it cleanly divides the 11,025kHz samplerate. Oddly
enough, the lip sync does not suffer from this at all; even at 11fps
the intro video keeps audio synced exactly to the video.

This also reverts commit 56f527bdd3b66937fd33a0a473b0ca742f4db5ce as
it is completely unnecessary.