sandec: implement GOST
GOST is a 12byte (RA1) / 6byte (RA2+) chunk with a command to
identify the flip planes, and additional x/y offsets (like FTCH).
These are then applied to the previous FOBJ.  The flip flags are
only ever implemented in codec1, so I've added quick hacks for
codec1 flipped x, y, and both.

Now, in the RA1 FNFINAL.ANM file, both sides of the audience clap,
not just the left one.